pyproject.toml
[tool.poetry]
name = "discounts"
version = "0.1.0"
description = ""
authors = ["miko <islamimanhod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.6.0"
certifi = "2022.12.7"
charset-normalizer = "3.1.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
Django = "3.2.5"
django-environ = "0.10.0"
django-filter = "2.4.0"
djangorestframework = "3.12.4"
idna = "3.4"
inflection = "0.5.1"
itypes = "1.2.0"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.2"
packaging = "23.1"
Pillow = "8.3.1"
psycopg2-binary = "2.9.6"
pytz = "2023.3"
requests = "2.28.2"
"ruamel.yaml" = "0.17.21"
"ruamel.yaml.clib" = "0.2.7"
sqlparse = "0.4.4"
uritemplate = "4.1.1"
urllib3 = "1.26.15"
django-redis = "^5.2.0"
huey = "^2.4.5"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"

[tool.poetry.dev-dependencies]
cfgv = "3.3.1"
charset-normalizer = "3.1.0"
click = "8.1.3"
distlib = "0.3.6"
flake8 = "6.0.0"
identify = "2.5.22"
idna = "3.4"
isort = "5.12.0"
mccabe = "0.7.0"
mypy-extensions = "1.0.0"
nodeenv = "1.7.0"
pathspec = "0.11.1"
platformdirs = "3.2.0"
pre-commit = "3.2.2"
"ruamel.yaml.clib" = "0.2.7"
sqlparse = "0.4.4"
tomli = "2.0.1"
typing-extensions = "4.5.0"
tzlocal = "2.1"
uritemplate = "4.1.1"
urllib3 = "1.26.15"
django-debug-toolbar = "4.0.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "configs.settings"
python_files = ["test_*.py"]
